refactor: split app into modular files
